# Support filtering with almost any number of paths