--- a/.changeset/nine-countries-join.md
+++ b/.changeset/nine-countries-join.md
@@ -1,0 +1,5 @@
+---
+"@itwin/presentation-hierarchies": patch
+---
+
+Changed hierarchy filtering, it now works with any number of filter paths.