--- a/packages/hierarchies/src/test/HierarchyFiltering.test.ts
+++ b/packages/hierarchies/src/test/HierarchyFiltering.test.ts
@@ -1,0 +1,44 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import { expect } from "chai";
+import { HierarchyFilteringPath, HierarchyFilteringPathOptions } from "../hierarchies/HierarchyFiltering";
+
+describe("HierarchyFilteringPath", () => {
+  describe("mergeOptions", () => {
+    const optionsInOrderOfPriority: Array<HierarchyFilteringPathOptions | undefined> = [
+      { autoExpand: true },
+      {
+        autoExpand: {
+          key: {
+            type: "label-grouping",
+            label: "",
+          },
+          depth: 2,
+        },
+      },
+      {
+        autoExpand: {
+          key: {
+            type: "label-grouping",
+            label: "",
+          },
+          depth: 1,
+        },
+      },
+      { autoExpand: false },
+      undefined,
+    ];
+    it("returns correct result for different types of identifiers", () => {
+      for (let i = 0; i < optionsInOrderOfPriority.length; ++i) {
+        for (let j = 0; j < optionsInOrderOfPriority.length; ++j) {
+          expect(HierarchyFilteringPath.mergeOptions(optionsInOrderOfPriority[i], optionsInOrderOfPriority[j])).to.deep.eq(
+            i < j ? optionsInOrderOfPriority[i] : optionsInOrderOfPriority[j],
+          );
+        }
+      }
+    });
+  });
+});