Closes https://github.com/iTwin/presentation/issues/722

Performance benchmark says that the test is about 1sec slower than with previous implementation. That is because new implementation puts 1k nodes under a single node (previously it was 500). When putting more and more nodes under a single node the filtering gets slower and slower. Ran the tests 10 times and got these results

|     | Previous implementation      |   New implementation with 500 paths    | New implementation with 1k paths      |
| ------------- | ------------- | ------------- | ------------- |
| Min | 16426.47 ms | 14686.84 ms  | 16929.78 ms |
| Max | 17432.14 ms | 16331.06 ms | 18527.86 ms |
| Average | 16950.287 ms | 15416.8 ms | 17756.909 ms |